Estimate seroprevalence and FOI from a fixed mixture model

Estimate age-specific seroprevalence and FOI given a fitted mixture model (generated by [serosv::mixture_model()])

Usage

estimate_from_mixture(
  age,
  antibody_level,
  threshold_status = NULL,
  mixture_model,
  s = "ps",
  sp = 83,
  monotonize = TRUE
)

Arguments

age

vector of age

antibody_level

vector of the corresponding raw antibody level

threshold_status

sero status using threshold approach in line listing (optional, for visualization and comparison only)

mixture_model

mixture_model object generated by serosv::mixture_model()

s

smoothing basis used to fit antibody level

sp

smoothing parameter

monotonize

whether to monotonize seroprevalence (default to TRUE)

Value

a list of class estimated_from_mixture with the following items

df

the dataframe used for fitting the model

info

a fitted "gam" model for mu(a)

sp

seroprevalence

foi

force of infection

threshold_status

serostatus using threshold method only if provided

Details

Antibody level (denoted \(Z\)) is modeled using a 2-component Gaussian mixture model. Each component \(Z_j\) (\(j \in \{I, S\}\)) represents the antibody level of the latent Infected and Susceptible sub-populations, following density \(f_j(z_j|\theta_j)\)

Let \(\pi_{\text{TRUE}}(a)\) denotes the age-dependent mixing probability (i.e., the true prevalence), the density of the mixture is formulated as

$$f(z|z_I, z_S,a) = (1-\pi_{\text{TRUE}}(a))f_S(z_S|\theta_S)+\pi_{\text{TRUE}}(a)f_I(z_I|\theta_I)$$

The mean \(E(Z|a)\) thus equals $$\mu(a) = (1-\pi_{\text{TRUE}}(a))\mu_S+\pi_{\text{TRUE}}(a)\mu_I$$

From which true prevalence can be computed as $$\pi_{\text{TRUE}}(a) = \frac{\mu(a) - \mu_S}{\mu_I - \mu_S}$$

And FOI can then be inferred as $$\lambda_{TRUE} = \frac{\mu'(a)}{\mu_I - \mu(a)}$$

Function [serosv::mixture_model()] fits antibody level data to \(f_S(z_S|\theta_S)\) and \(f_I(z_I|\theta_I)\)

Function [serosv::estimate_mixture()] will then estimate age-specific antibody level \(\mu(a)\) and infer the estimation for \(\pi_{\text{TRUE}}(a)\) and \(\lambda_{TRUE}\)

Refer to section 11.3. of the the book by Hens et al. (2012) for further details.

References

Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme, and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on Serological and Social Contact Data: A Modern Statistical Perspective. tatistics for Biology and Health. Springer New York. doi:10.1007/978-1-4614-4072-7 .

See also