serosv: model infectious disease parameters

An easy-to-use and efficient tool to estimate infectious diseases parameters using serological data. Implemented models include SIR models (basic_sir_model(), static_sir_model(), mseir_model(), sir_subpops_model()), parametric models (polynomial_model(), fp_model()), nonparametric models (lp_model()), semiparametric models (penalized_splines_model()), hierarchical models (hierarchical_bayesian_model()). The package is based on the book "Modeling Infectious Disease Parameters Based on Serological and Social Contact Data: A Modern Statistical Perspective" (Hens, Niel & Shkedy, Ziv & Aerts, Marc & Faes, Christel & Damme, Pierre & Beutels, Philippe., 2013) doi:10.1007/978-1-4614-4072-7 .

See also

Author

Maintainer: Anh Phan Truong Quynh anhptq@oucru.org (ORCID)

Authors:

Nguyen Pham Nguyen The nguyenpnt@oucru.org (ORCID)

Long Bui Thanh

Tuyen Huynh tuyenhn@oucru.org

Thinh Ong thinhop@oucru.org (ORCID)

Marc Choisy mchoisy@oucru.org (ORCID)