SIR static model (age-heterogeneous, endemic equilibrium)

Simulate transmission model with constant force of infection and consists of 3 compartments: susceptible (S), infected (I), recovered (R)

Usage

sir_static_model(a, state, parameters)

Arguments

a

age sequence.

state

the initial state of the system. A named vector with the following

s: initial susceptible proportion

i: initial infected proportion

r: initial recovered proportion

parameters

the model's parameter. A named vector with the following

lambda: natural death rate

nu: recovery rate

Value

list of class sir_static_model with the following items

parameters

list of parameters used for fitting the model

output

matrix of proportion for each compartment over time

Details

Follow the SIR model at endemic state described in the book by Hens et al. (section 3.2.2.)

Assumptions:

- Time homogeneity

- Age heterogeneity

- Constant force of infection

The model is described by a system of 3 differential equations

$$ \begin{cases} \frac{ds(a)}{da} = -\lambda s(a) \\ \frac{di(a)}{da} = \lambda s(a) - \nu i(a) \\ \frac{dr(a)}{da} = \nu i(a) \end{cases} $$

Where:

- \(s(a), i(a), r(a)\) are proportion of susceptible, infected, recovered population of age group \(a\) respectively

- \(\lambda\) is the force of infection

- \(\nu\) is the recovery rate

References

Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme, and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on Serological and Social Contact Data: A Modern Statistical Perspective. tatistics for Biology and Health. Springer New York. doi:10.1007/978-1-4614-4072-7 .

Examples

state <- c(s=0.99,i=0.01,r=0)
parameters <- c(
  lambda = 0.05,
  nu=1/(14/365) # 2 weeks to recover
)
ages<-seq(0, 90, by=0.01)
model = sir_static_model(ages, state, parameters)
model
#> $parameters
#>   lambda       nu 
#>  0.05000 26.07143 
#> 
#> $output
#>     time         s           i           r
#> 1   0.00 0.9900000 0.010000000 0.000000000
#> 2   0.01 0.9895051 0.008139822 0.002355054
#> 3   0.02 0.9890105 0.006708036 0.004281469
#> 4   0.03 0.9885161 0.005603950 0.005879937
#> 5   0.04 0.9880220 0.004752794 0.007225227
#> 6   0.05 0.9875281 0.004096558 0.008375351
#> 7   0.06 0.9870345 0.003590872 0.009374678
#> 8   0.07 0.9865411 0.003201102 0.010257841
#> 9   0.08 0.9860479 0.002900525 0.011051565
#> 10  0.09 0.9855550 0.002668724 0.011776268
#> 11  0.10 0.9850624 0.002489909 0.012447737
#> 12  0.11 0.9845699 0.002351916 0.013078137
#> 13  0.12 0.9840778 0.002245378 0.013676837
#> 14  0.13 0.9835859 0.002163076 0.014251056
#> 15  0.14 0.9830942 0.002099446 0.014806355
#> 16  0.15 0.9826028 0.002050204 0.015347022
#> 17  0.16 0.9821116 0.002012047 0.015876357
#> 18  0.17 0.9816207 0.001982432 0.016396906
#> 19  0.18 0.9811300 0.001959397 0.016910628
#> 20  0.19 0.9806395 0.001941438 0.017419030
#> 21  0.20 0.9801493 0.001927389 0.017923276
#> 22  0.21 0.9796594 0.001916345 0.018424271
#> 23  0.22 0.9791697 0.001907612 0.018922712
#> 24  0.23 0.9786802 0.001900660 0.019419127
#> 25  0.24 0.9781910 0.001895079 0.019913926
#> 26  0.25 0.9777020 0.001890571 0.020407406
#> 27  0.26 0.9772133 0.001886886 0.020899820
#> 28  0.27 0.9767248 0.001883834 0.021391357
#> 29  0.28 0.9762366 0.001881269 0.021882162
#> 30  0.29 0.9757486 0.001879079 0.022372348
#> 31  0.30 0.9752608 0.001877178 0.022862001
#> 32  0.31 0.9747733 0.001875500 0.023351189
#> 33  0.32 0.9742860 0.001873993 0.023839961
#> 34  0.33 0.9737990 0.001872617 0.024328357
#> 35  0.34 0.9733122 0.001871344 0.024816409
#> 36  0.35 0.9728257 0.001870148 0.025304138
#> 37  0.36 0.9723394 0.001869014 0.025791565
#> 38  0.37 0.9718534 0.001867925 0.026278701
#> 39  0.38 0.9713676 0.001866873 0.026765558
#> 40  0.39 0.9708820 0.001865849 0.027252145
#> 41  0.40 0.9703967 0.001864846 0.027738468
#> 42  0.41 0.9699116 0.001863860 0.028224531
#> 43  0.42 0.9694268 0.001862886 0.028710339
#> 44  0.43 0.9689422 0.001861923 0.029195894
#> 45  0.44 0.9684578 0.001860968 0.029681200
#> 46  0.45 0.9679737 0.001860019 0.030166257
#> 47  0.46 0.9674899 0.001859074 0.030651067
#> 48  0.47 0.9670062 0.001858134 0.031135632
#> 49  0.48 0.9665229 0.001857196 0.031619952
#> 50  0.49 0.9660397 0.001856261 0.032104028
#> 51  0.50 0.9655568 0.001855328 0.032587860
#> 52  0.51 0.9650742 0.001854396 0.033071449
#> 53  0.52 0.9645917 0.001853466 0.033554795
#> 54  0.53 0.9641096 0.001852538 0.034037900
#> 55  0.54 0.9636276 0.001851610 0.034520762
#> 56  0.55 0.9631459 0.001850683 0.035003382
#> 57  0.56 0.9626645 0.001849757 0.035485761
#> 58  0.57 0.9621833 0.001848831 0.035967898
#> 59  0.58 0.9617023 0.001847906 0.036449794
#> 60  0.59 0.9612216 0.001846982 0.036931450
#> 61  0.60 0.9607411 0.001846059 0.037412864
#> 62  0.61 0.9602608 0.001845136 0.037894037
#> 63  0.62 0.9597808 0.001844213 0.038374970
#> 64  0.63 0.9593010 0.001843291 0.038855663
#> 65  0.64 0.9588215 0.001842369 0.039336115
#> 66  0.65 0.9583422 0.001841448 0.039816327
#> 67  0.66 0.9578632 0.001840528 0.040296299
#> 68  0.67 0.9573844 0.001839608 0.040776031
#> 69  0.68 0.9569058 0.001838688 0.041255523
#> 70  0.69 0.9564275 0.001837769 0.041734776
#> 71  0.70 0.9559494 0.001836850 0.042213789
#> 72  0.71 0.9554715 0.001835932 0.042692562
#> 73  0.72 0.9549939 0.001835014 0.043171096
#> 74  0.73 0.9545165 0.001834097 0.043649391
#> 75  0.74 0.9540394 0.001833180 0.044127447
#> 76  0.75 0.9535625 0.001832264 0.044605264
#> 77  0.76 0.9530858 0.001831348 0.045082842
#> 78  0.77 0.9526094 0.001830433 0.045560181
#> 79  0.78 0.9521332 0.001829518 0.046037282
#> 80  0.79 0.9516573 0.001828603 0.046514144
#> 81  0.80 0.9511815 0.001827689 0.046990768
#> 82  0.81 0.9507061 0.001826775 0.047467153
#> 83  0.82 0.9502308 0.001825862 0.047943300
#> 84  0.83 0.9497558 0.001824949 0.048419210
#> 85  0.84 0.9492811 0.001824037 0.048894881
#> 86  0.85 0.9488066 0.001823125 0.049370315
#> 87  0.86 0.9483323 0.001822214 0.049845511
#> 88  0.87 0.9478582 0.001821303 0.050320470
#> 89  0.88 0.9473844 0.001820393 0.050795191
#> 90  0.89 0.9469108 0.001819483 0.051269675
#> 91  0.90 0.9464375 0.001818573 0.051743921
#> 92  0.91 0.9459644 0.001817664 0.052217931
#> 93  0.92 0.9454915 0.001816756 0.052691703
#> 94  0.93 0.9450189 0.001815847 0.053165239
#> 95  0.94 0.9445465 0.001814940 0.053638538
#> 96  0.95 0.9440744 0.001814033 0.054111601
#> 97  0.96 0.9436024 0.001813126 0.054584427
#> 98  0.97 0.9431308 0.001812219 0.055057016
#> 99  0.98 0.9426593 0.001811314 0.055529370
#> 100 0.99 0.9421881 0.001810408 0.056001487
#> 101 1.00 0.9417171 0.001809503 0.056473368
#> 102 1.01 0.9412464 0.001808599 0.056945014
#> 103 1.02 0.9407759 0.001807695 0.057416423
#> 104 1.03 0.9403056 0.001806791 0.057887597
#> 105 1.04 0.9398356 0.001805888 0.058358536
#> 106 1.05 0.9393658 0.001804985 0.058829239
#> 107 1.06 0.9388962 0.001804083 0.059299707
#> 108 1.07 0.9384269 0.001803181 0.059769939
#> 109 1.08 0.9379578 0.001802280 0.060239937
#> 110 1.09 0.9374889 0.001801379 0.060709699
#> 111 1.10 0.9370203 0.001800478 0.061179227
#> 112 1.11 0.9365519 0.001799578 0.061648520
#> 113 1.12 0.9360837 0.001798679 0.062117579
#> 114 1.13 0.9356158 0.001797780 0.062586403
#> 115 1.14 0.9351481 0.001796881 0.063054992
#> 116 1.15 0.9346807 0.001795983 0.063523348
#> 117 1.16 0.9342134 0.001795085 0.063991469
#> 118 1.17 0.9337465 0.001794188 0.064459356
#> 119 1.18 0.9332797 0.001793291 0.064927010
#> 120 1.19 0.9328132 0.001792394 0.065394430
#> 121 1.20 0.9323469 0.001791498 0.065861615
#> 122 1.21 0.9318808 0.001790603 0.066328568
#> 123 1.22 0.9314150 0.001789708 0.066795287
#> 124 1.23 0.9309494 0.001788813 0.067261773
#> 125 1.24 0.9304841 0.001787919 0.067728025
#> 126 1.25 0.9300189 0.001787025 0.068194045
#> 127 1.26 0.9295540 0.001786132 0.068659831
#> 128 1.27 0.9290894 0.001785239 0.069125385
#> 129 1.28 0.9286249 0.001784347 0.069590706
#> 130 1.29 0.9281608 0.001783455 0.070055794
#> 131 1.30 0.9276968 0.001782563 0.070520650
#> 132 1.31 0.9272331 0.001781672 0.070985274
#> 133 1.32 0.9267696 0.001780781 0.071449665
#> 134 1.33 0.9263063 0.001779891 0.071913824
#> 135 1.34 0.9258432 0.001779002 0.072377751
#> 136 1.35 0.9253804 0.001778112 0.072841447
#> 137 1.36 0.9249179 0.001777223 0.073304910
#> 138 1.37 0.9244555 0.001776335 0.073768142
#> 139 1.38 0.9239934 0.001775447 0.074231142
#> 140 1.39 0.9235315 0.001774560 0.074693911
#> 141 1.40 0.9230699 0.001773673 0.075156448
#> 142 1.41 0.9226085 0.001772786 0.075618754
#> 143 1.42 0.9221473 0.001771900 0.076080829
#> 144 1.43 0.9216863 0.001771014 0.076542674
#> 145 1.44 0.9212256 0.001770129 0.077004287
#> 146 1.45 0.9207651 0.001769244 0.077465669
#> 147 1.46 0.9203048 0.001768360 0.077926821
#> 148 1.47 0.9198448 0.001767476 0.078387743
#> 149 1.48 0.9193850 0.001766592 0.078848434
#> 150 1.49 0.9189254 0.001765709 0.079308894
#> 151 1.50 0.9184660 0.001764826 0.079769125
#> 152 1.51 0.9180069 0.001763944 0.080229125
#> 153 1.52 0.9175480 0.001763062 0.080688896
#> 154 1.53 0.9170894 0.001762181 0.081148436
#> 155 1.54 0.9166310 0.001761300 0.081607747
#> 156 1.55 0.9161728 0.001760420 0.082066829
#> 157 1.56 0.9157148 0.001759540 0.082525681
#> 158 1.57 0.9152570 0.001758660 0.082984303
#> 159 1.58 0.9147995 0.001757781 0.083442696
#> 160 1.59 0.9143422 0.001756902 0.083900860
#> 161 1.60 0.9138852 0.001756024 0.084358796
#> 162 1.61 0.9134284 0.001755146 0.084816502
#> 163 1.62 0.9129718 0.001754269 0.085273979
#> 164 1.63 0.9125154 0.001753392 0.085731228
#> 165 1.64 0.9120592 0.001752516 0.086188248
#> 166 1.65 0.9116033 0.001751640 0.086645040
#> 167 1.66 0.9111476 0.001750764 0.087101603
#> 168 1.67 0.9106922 0.001749889 0.087557938
#> 169 1.68 0.9102369 0.001749014 0.088014045
#> 170 1.69 0.9097819 0.001748140 0.088469924
#> 171 1.70 0.9093272 0.001747266 0.088925575
#> 172 1.71 0.9088726 0.001746393 0.089380999
#> 173 1.72 0.9084183 0.001745520 0.089836194
#> 174 1.73 0.9079642 0.001744647 0.090291163
#> 175 1.74 0.9075103 0.001743775 0.090745903
#> 176 1.75 0.9070567 0.001742903 0.091200417
#> 177 1.76 0.9066033 0.001742032 0.091654703
#> 178 1.77 0.9061501 0.001741161 0.092108762
#> 179 1.78 0.9056971 0.001740291 0.092562594
#> 180 1.79 0.9052444 0.001739421 0.093016200
#> 181 1.80 0.9047919 0.001738551 0.093469578
#> 182 1.81 0.9043396 0.001737682 0.093922730
#> 183 1.82 0.9038875 0.001736814 0.094375655
#> 184 1.83 0.9034357 0.001735946 0.094828354
#> 185 1.84 0.9029841 0.001735078 0.095280827
#> 186 1.85 0.9025327 0.001734211 0.095733074
#> 187 1.86 0.9020816 0.001733344 0.096185094
#> 188 1.87 0.9016306 0.001732477 0.096636889
#> 189 1.88 0.9011799 0.001731611 0.097088457
#> 190 1.89 0.9007295 0.001730746 0.097539800
#> 191 1.90 0.9002792 0.001729880 0.097990918
#> 192 1.91 0.8998292 0.001729016 0.098441809
#> 193 1.92 0.8993794 0.001728151 0.098892476
#> 194 1.93 0.8989298 0.001727288 0.099342917
#> 195 1.94 0.8984804 0.001726424 0.099793133
#> 196 1.95 0.8980313 0.001725561 0.100243124
#> 197 1.96 0.8975824 0.001724699 0.100692890
#> 198 1.97 0.8971337 0.001723836 0.101142431
#> 199 1.98 0.8966853 0.001722975 0.101591748
#> 200 1.99 0.8962370 0.001722113 0.102040840
#> 201 2.00 0.8957890 0.001721253 0.102489707
#> 202 2.01 0.8953413 0.001720392 0.102938350
#> 203 2.02 0.8948937 0.001719532 0.103386769
#> 204 2.03 0.8944464 0.001718673 0.103834963
#> 205 2.04 0.8939993 0.001717814 0.104282934
#> 206 2.05 0.8935524 0.001716955 0.104730680
#> 207 2.06 0.8931057 0.001716097 0.105178203
#> 208 2.07 0.8926593 0.001715239 0.105625502
#> 209 2.08 0.8922130 0.001714381 0.106072578
#> 210 2.09 0.8917670 0.001713524 0.106519430
#> 211 2.10 0.8913213 0.001712668 0.106966058
#> 212 2.11 0.8908757 0.001711812 0.107412464
#> 213 2.12 0.8904304 0.001710956 0.107858646
#> 214 2.13 0.8899853 0.001710101 0.108304605
#> 215 2.14 0.8895404 0.001709246 0.108750341
#> 216 2.15 0.8890958 0.001708392 0.109195855
#> 217 2.16 0.8886513 0.001707538 0.109641146
#> 218 2.17 0.8882071 0.001706684 0.110086214
#> 219 2.18 0.8877631 0.001705831 0.110531060
#> 220 2.19 0.8873193 0.001704978 0.110975683
#> 221 2.20 0.8868758 0.001704126 0.111420084
#> 222 2.21 0.8864325 0.001703274 0.111864263
#> 223 2.22 0.8859894 0.001702423 0.112308220
#> 224 2.23 0.8855465 0.001701572 0.112751955
#> 225 2.24 0.8851038 0.001700721 0.113195468
#> 226 2.25 0.8846614 0.001699871 0.113638759
#> 227 2.26 0.8842191 0.001699021 0.114081829
#> 228 2.27 0.8837772 0.001698172 0.114524678
#> 229 2.28 0.8833354 0.001697323 0.114967305
#> 230 2.29 0.8828938 0.001696475 0.115409710
#> 231 2.30 0.8824525 0.001695627 0.115851895
#> 232 2.31 0.8820114 0.001694779 0.116293858
#> 233 2.32 0.8815705 0.001693932 0.116735601
#> 234 2.33 0.8811298 0.001693085 0.117177123
#> 235 2.34 0.8806893 0.001692239 0.117618424
#> 236 2.35 0.8802491 0.001691393 0.118059505
#> 237 2.36 0.8798091 0.001690547 0.118500365
#> 238 2.37 0.8793693 0.001689702 0.118941004
#> 239 2.38 0.8789297 0.001688858 0.119381424
#> 240 2.39 0.8784904 0.001688013 0.119821623
#> 241 2.40 0.8780512 0.001687170 0.120261602
#> 242 2.41 0.8776123 0.001686326 0.120701361
#> 243 2.42 0.8771736 0.001685483 0.121140901
#> 244 2.43 0.8767351 0.001684641 0.121580221
#> 245 2.44 0.8762969 0.001683799 0.122019321
#> 246 2.45 0.8758588 0.001682957 0.122458201
#> 247 2.46 0.8754210 0.001682116 0.122896863
#> 248 2.47 0.8749834 0.001681275 0.123335305
#> 249 2.48 0.8745460 0.001680434 0.123773527
#> 250 2.49 0.8741089 0.001679594 0.124211531
#>  [ reached 'max' / getOption("max.print") -- omitted 8751 rows ]
#> 
#> attr(,"class")
#> [1] "sir_static_model"