SIR Model with Interacting Subpopulations

An extension of the basic SIR model that incorporates interaction between sub-populations

Usage

sir_subpops_model(times, state, parameters)

Arguments

times

time sequence.

state

the initial state of the model. A named vector with the following

s: initial susceptible proportion

i: initial infected proportion

r: initial recovered proportion

parameters

the parameters of the model. A named vector with the following

mu: natural death rate (1/L).

beta: the WAIFW matrix, with dimensions [K, K].

nu: recovery rate

Value

list of class sir_subpops_model with the following items

parameters

list of parameters used for fitting the model

output

matrix of proportion for each compartment over time

Details

Follow the SIR model with sub populations described in the book by Hens et al. (section 3.5.1.)

With K subpopulations, the WAIFW matrix or mixing matrix is given by

$$ C = \begin{bmatrix} \beta_{11} & \beta_{12} & ... & \beta_{1K} \\ \beta_{21} & \beta_{22} & ... & \beta_{2K} \\ \vdots & \vdots & ... & \vdots \\ \beta_{K1} & \beta_{K2} & ... & \beta_{KK} \\ \end{bmatrix} $$

And the \(i^{th}\) sub population is described by the following system of differential equations $$ \begin{cases} \frac{dS_i(t)}{dt} = -(\sum^K_{j=1}\beta_{ij}I_j(t)) S_i(t) + N_i\mu_i - \mu_i S_i(t) \\ \frac{dI_i(t)}{dt} = (\sum^K_{j=1}\beta_{ij}I_j(t)) S_i(t) - (\nu_i + \mu_i) I_i(t) \\ \frac{dR_i(t)}{dt} = \nu_i I_i(t) - \mu_i R_i(t) \end{cases} $$

References

Hens, Niel, Ziv Shkedy, Marc Aerts, Christel Faes, Pierre Van Damme, and Philippe Beutels. 2012. Modeling Infectious Disease Parameters Based on Serological and Social Contact Data: A Modern Statistical Perspective. tatistics for Biology and Health. Springer New York. doi:10.1007/978-1-4614-4072-7 .

Examples

# \donttest{
state <- c(
  s = c(0.8, 0.8),
  i = c(0.2, 0.2),
  r = c(  0,   0)
)
beta_matrix <- matrix(
  c(0.05, 0.00,
  0.00, 0.05),
  2
)
parameters <- list(
  beta = beta_matrix,
  nu = c(1/30, 1/30),
  mu = 0.001
)
times<-seq(0,10000,by=0.5)
model <- sir_subpops_model(times, state, parameters)
model
#> $parameters
#> $parameters$beta
#>      [,1] [,2]
#> [1,] 0.05 0.00
#> [2,] 0.00 0.05
#> 
#> $parameters$nu
#> [1] 0.03333333 0.03333333
#> 
#> $parameters$mu
#> [1] 0.001
#> 
#> $parameters$k
#> [1] 2
#> 
#> 
#> $output
#>     time        s1        s2        i1        i2          r1          r2
#> 1    0.0 0.8000000 0.8000000 0.2000000 0.2000000 0.000000000 0.000000000
#> 2    0.5 0.7961052 0.7961052 0.2005577 0.2005577 0.003337146 0.003337146
#> 3    1.0 0.7922208 0.7922208 0.2010974 0.2010974 0.006681767 0.006681767
#> 4    1.5 0.7883475 0.7883475 0.2016190 0.2016190 0.010033558 0.010033558
#> 5    2.0 0.7844854 0.7844854 0.2021224 0.2021224 0.013392251 0.013392251
#> 6    2.5 0.7806350 0.7806350 0.2026075 0.2026075 0.016757489 0.016757489
#> 7    3.0 0.7767967 0.7767967 0.2030743 0.2030743 0.020128975 0.020128975
#> 8    3.5 0.7729709 0.7729709 0.2035227 0.2035227 0.023506400 0.023506400
#> 9    4.0 0.7691580 0.7691580 0.2039525 0.2039525 0.026889453 0.026889453
#> 10   4.5 0.7653583 0.7653583 0.2043639 0.2043639 0.030277824 0.030277824
#> 11   5.0 0.7615722 0.7615722 0.2047566 0.2047566 0.033671200 0.033671200
#> 12   5.5 0.7578000 0.7578000 0.2051307 0.2051307 0.037069269 0.037069269
#> 13   6.0 0.7540421 0.7540421 0.2054862 0.2054862 0.040471717 0.040471717
#> 14   6.5 0.7502988 0.7502988 0.2058229 0.2058229 0.043878231 0.043878231
#> 15   7.0 0.7465705 0.7465705 0.2061410 0.2061410 0.047288499 0.047288499
#> 16   7.5 0.7428574 0.7428574 0.2064404 0.2064404 0.050702206 0.050702206
#> 17   8.0 0.7391599 0.7391599 0.2067210 0.2067210 0.054119038 0.054119038
#> 18   8.5 0.7354784 0.7354784 0.2069829 0.2069829 0.057538683 0.057538683
#> 19   9.0 0.7318130 0.7318130 0.2072262 0.2072262 0.060960827 0.060960827
#> 20   9.5 0.7281641 0.7281641 0.2074507 0.2074507 0.064385157 0.064385157
#> 21  10.0 0.7245320 0.7245320 0.2076566 0.2076566 0.067811362 0.067811362
#> 22  10.5 0.7209170 0.7209170 0.2078439 0.2078439 0.071239129 0.071239129
#> 23  11.0 0.7173193 0.7173193 0.2080126 0.2080126 0.074668149 0.074668149
#> 24  11.5 0.7137392 0.7137392 0.2081627 0.2081627 0.078098110 0.078098110
#> 25  12.0 0.7101769 0.7101769 0.2082944 0.2082944 0.081528705 0.081528705
#> 26  12.5 0.7066328 0.7066328 0.2084076 0.2084076 0.084959625 0.084959625
#> 27  13.0 0.7031070 0.7031070 0.2085024 0.2085024 0.088390563 0.088390563
#> 28  13.5 0.6995998 0.6995998 0.2085790 0.2085790 0.091821214 0.091821214
#> 29  14.0 0.6961114 0.6961114 0.2086373 0.2086373 0.095251274 0.095251274
#> 30  14.5 0.6926420 0.6926420 0.2086776 0.2086776 0.098680441 0.098680441
#> 31  15.0 0.6891919 0.6891919 0.2086997 0.2086997 0.102108412 0.102108412
#> 32  15.5 0.6857612 0.6857612 0.2087039 0.2087039 0.105534890 0.105534890
#> 33  16.0 0.6823501 0.6823501 0.2086903 0.2086903 0.108959577 0.108959577
#> 34  16.5 0.6789589 0.6789589 0.2086589 0.2086589 0.112382176 0.112382176
#> 35  17.0 0.6755877 0.6755877 0.2086099 0.2086099 0.115802395 0.115802395
#> 36  17.5 0.6722366 0.6722366 0.2085434 0.2085434 0.119219941 0.119219941
#> 37  18.0 0.6689060 0.6689060 0.2084595 0.2084595 0.122634526 0.122634526
#> 38  18.5 0.6655958 0.6655958 0.2083583 0.2083583 0.126045862 0.126045862
#> 39  19.0 0.6623063 0.6623063 0.2082400 0.2082400 0.129453664 0.129453664
#> 40  19.5 0.6590376 0.6590376 0.2081047 0.2081047 0.132857649 0.132857649
#> 41  20.0 0.6557899 0.6557899 0.2079526 0.2079526 0.136257538 0.136257538
#> 42  20.5 0.6525632 0.6525632 0.2077838 0.2077838 0.139653053 0.139653053
#> 43  21.0 0.6493577 0.6493577 0.2075983 0.2075983 0.143043919 0.143043919
#> 44  21.5 0.6461736 0.6461736 0.2073965 0.2073965 0.146429864 0.146429864
#> 45  22.0 0.6430109 0.6430109 0.2071785 0.2071785 0.149810618 0.149810618
#> 46  22.5 0.6398697 0.6398697 0.2069444 0.2069444 0.153185914 0.153185914
#> 47  23.0 0.6367502 0.6367502 0.2066943 0.2066943 0.156555490 0.156555490
#> 48  23.5 0.6336524 0.6336524 0.2064286 0.2064286 0.159919084 0.159919084
#> 49  24.0 0.6305764 0.6305764 0.2061472 0.2061472 0.163276437 0.163276437
#> 50  24.5 0.6275222 0.6275222 0.2058505 0.2058505 0.166627296 0.166627296
#> 51  25.0 0.6244901 0.6244901 0.2055385 0.2055385 0.169971409 0.169971409
#> 52  25.5 0.6214799 0.6214799 0.2052115 0.2052115 0.173308527 0.173308527
#> 53  26.0 0.6184919 0.6184919 0.2048697 0.2048697 0.176638404 0.176638404
#> 54  26.5 0.6155259 0.6155259 0.2045133 0.2045133 0.179960799 0.179960799
#> 55  27.0 0.6125822 0.6125822 0.2041423 0.2041423 0.183275473 0.183275473
#> 56  27.5 0.6096607 0.6096607 0.2037571 0.2037571 0.186582190 0.186582190
#> 57  28.0 0.6067614 0.6067614 0.2033579 0.2033579 0.189880719 0.189880719
#> 58  28.5 0.6038844 0.6038844 0.2029447 0.2029447 0.193170830 0.193170830
#> 59  29.0 0.6010298 0.6010298 0.2025179 0.2025179 0.196452298 0.196452298
#> 60  29.5 0.5981974 0.5981974 0.2020777 0.2020777 0.199724902 0.199724902
#> 61  30.0 0.5953875 0.5953875 0.2016241 0.2016241 0.202988423 0.202988423
#> 62  30.5 0.5925998 0.5925998 0.2011575 0.2011575 0.206242646 0.206242646
#> 63  31.0 0.5898346 0.5898346 0.2006781 0.2006781 0.209487361 0.209487361
#> 64  31.5 0.5870917 0.5870917 0.2001860 0.2001860 0.212722359 0.212722359
#> 65  32.0 0.5843711 0.5843711 0.1996815 0.1996815 0.215947437 0.215947437
#> 66  32.5 0.5816729 0.5816729 0.1991647 0.1991647 0.219162394 0.219162394
#> 67  33.0 0.5789970 0.5789970 0.1986360 0.1986360 0.222367034 0.222367034
#> 68  33.5 0.5763434 0.5763434 0.1980954 0.1980954 0.225561163 0.225561163
#> 69  34.0 0.5737121 0.5737121 0.1975433 0.1975433 0.228744592 0.228744592
#> 70  34.5 0.5711030 0.5711030 0.1969798 0.1969798 0.231917135 0.231917135
#> 71  35.0 0.5685162 0.5685162 0.1964052 0.1964052 0.235078610 0.235078610
#> 72  35.5 0.5659515 0.5659515 0.1958197 0.1958197 0.238228838 0.238228838
#> 73  36.0 0.5634090 0.5634090 0.1952234 0.1952234 0.241367645 0.241367645
#> 74  36.5 0.5608886 0.5608886 0.1946166 0.1946166 0.244494860 0.244494860
#> 75  37.0 0.5583902 0.5583902 0.1939995 0.1939995 0.247610314 0.247610314
#> 76  37.5 0.5559138 0.5559138 0.1933724 0.1933724 0.250713846 0.250713846
#> 77  38.0 0.5534593 0.5534593 0.1927354 0.1927354 0.253805294 0.253805294
#> 78  38.5 0.5510267 0.5510267 0.1920888 0.1920888 0.256884502 0.256884502
#> 79  39.0 0.5486160 0.5486160 0.1914327 0.1914327 0.259951318 0.259951318
#> 80  39.5 0.5462269 0.5462269 0.1907675 0.1907675 0.263005592 0.263005592
#> 81  40.0 0.5438596 0.5438596 0.1900932 0.1900932 0.266047180 0.266047180
#> 82  40.5 0.5415139 0.5415139 0.1894102 0.1894102 0.269075939 0.269075939
#> 83  41.0 0.5391896 0.5391896 0.1887186 0.1887186 0.272091732 0.272091732
#> 84  41.5 0.5368869 0.5368869 0.1880187 0.1880187 0.275094424 0.275094424
#> 85  42.0 0.5346055 0.5346055 0.1873106 0.1873106 0.278083884 0.278083884
#> 86  42.5 0.5323454 0.5323454 0.1865946 0.1865946 0.281059985 0.281059985
#> 87  43.0 0.5301065 0.5301065 0.1858709 0.1858709 0.284022604 0.284022604
#> 88  43.5 0.5278887 0.5278887 0.1851397 0.1851397 0.286971620 0.286971620
#> 89  44.0 0.5256919 0.5256919 0.1844012 0.1844012 0.289906917 0.289906917
#> 90  44.5 0.5235161 0.5235161 0.1836556 0.1836556 0.292828382 0.292828382
#> 91  45.0 0.5213610 0.5213610 0.1829031 0.1829031 0.295735905 0.295735905
#> 92  45.5 0.5192267 0.5192267 0.1821439 0.1821439 0.298629380 0.298629380
#> 93  46.0 0.5171131 0.5171131 0.1813782 0.1813782 0.301508705 0.301508705
#> 94  46.5 0.5150199 0.5150199 0.1806063 0.1806063 0.304373780 0.304373780
#> 95  47.0 0.5129472 0.5129472 0.1798283 0.1798283 0.307224510 0.307224510
#> 96  47.5 0.5108948 0.5108948 0.1790444 0.1790444 0.310060801 0.310060801
#> 97  48.0 0.5088626 0.5088626 0.1782548 0.1782548 0.312882565 0.312882565
#> 98  48.5 0.5068506 0.5068506 0.1774597 0.1774597 0.315689717 0.315689717
#> 99  49.0 0.5048585 0.5048585 0.1766593 0.1766593 0.318482173 0.318482173
#> 100 49.5 0.5028863 0.5028863 0.1758539 0.1758539 0.321259854 0.321259854
#> 101 50.0 0.5009338 0.5009338 0.1750435 0.1750435 0.324022684 0.324022684
#> 102 50.5 0.4990010 0.4990010 0.1742284 0.1742284 0.326770590 0.326770590
#> 103 51.0 0.4970878 0.4970878 0.1734087 0.1734087 0.329503503 0.329503503
#> 104 51.5 0.4951939 0.4951939 0.1725847 0.1725847 0.332221355 0.332221355
#> 105 52.0 0.4933194 0.4933194 0.1717565 0.1717565 0.334924084 0.334924084
#> 106 52.5 0.4914640 0.4914640 0.1709243 0.1709243 0.337611628 0.337611628
#> 107 53.0 0.4896277 0.4896277 0.1700883 0.1700883 0.340283931 0.340283931
#> 108 53.5 0.4878104 0.4878104 0.1692487 0.1692487 0.342940938 0.342940938
#> 109 54.0 0.4860118 0.4860118 0.1684056 0.1684056 0.345582596 0.345582596
#> 110 54.5 0.4842320 0.4842320 0.1675592 0.1675592 0.348208859 0.348208859
#> 111 55.0 0.4824707 0.4824707 0.1667096 0.1667096 0.350819678 0.350819678
#> 112 55.5 0.4807278 0.4807278 0.1658572 0.1658572 0.353415013 0.353415013
#> 113 56.0 0.4790033 0.4790033 0.1650019 0.1650019 0.355994822 0.355994822
#> 114 56.5 0.4772970 0.4772970 0.1641440 0.1641440 0.358559069 0.358559069
#> 115 57.0 0.4756087 0.4756087 0.1632836 0.1632836 0.361107718 0.361107718
#> 116 57.5 0.4739384 0.4739384 0.1624209 0.1624209 0.363640737 0.363640737
#> 117 58.0 0.4722858 0.4722858 0.1615561 0.1615561 0.366158098 0.366158098
#> 118 58.5 0.4706510 0.4706510 0.1606892 0.1606892 0.368659773 0.368659773
#> 119 59.0 0.4690337 0.4690337 0.1598205 0.1598205 0.371145738 0.371145738
#> 120 59.5 0.4674339 0.4674339 0.1589502 0.1589502 0.373615972 0.373615972
#> 121 60.0 0.4658513 0.4658513 0.1580782 0.1580782 0.376070455 0.376070455
#> 122 60.5 0.4642859 0.4642859 0.1572049 0.1572049 0.378509171 0.378509171
#> 123 61.0 0.4627376 0.4627376 0.1563303 0.1563303 0.380932105 0.380932105
#> 124 61.5 0.4612061 0.4612061 0.1554546 0.1554546 0.383339246 0.383339246
#> 125 62.0 0.4596915 0.4596915 0.1545779 0.1545779 0.385730583 0.385730583
#> 126 62.5 0.4581935 0.4581935 0.1537004 0.1537004 0.388106110 0.388106110
#> 127 63.0 0.4567120 0.4567120 0.1528221 0.1528221 0.390465822 0.390465822
#> 128 63.5 0.4552470 0.4552470 0.1519433 0.1519433 0.392809716 0.392809716
#> 129 64.0 0.4537982 0.4537982 0.1510641 0.1510641 0.395137790 0.395137790
#> 130 64.5 0.4523655 0.4523655 0.1501845 0.1501845 0.397450047 0.397450047
#> 131 65.0 0.4509488 0.4509488 0.1493047 0.1493047 0.399746490 0.399746490
#> 132 65.5 0.4495481 0.4495481 0.1484248 0.1484248 0.402027125 0.402027125
#> 133 66.0 0.4481630 0.4481630 0.1475450 0.1475450 0.404291960 0.404291960
#> 134 66.5 0.4467936 0.4467936 0.1466654 0.1466654 0.406541004 0.406541004
#> 135 67.0 0.4454397 0.4454397 0.1457860 0.1457860 0.408774268 0.408774268
#> 136 67.5 0.4441012 0.4441012 0.1449070 0.1449070 0.410991767 0.410991767
#> 137 68.0 0.4427779 0.4427779 0.1440286 0.1440286 0.413193516 0.413193516
#> 138 68.5 0.4414698 0.4414698 0.1431507 0.1431507 0.415379531 0.415379531
#> 139 69.0 0.4401766 0.4401766 0.1422736 0.1422736 0.417549833 0.417549833
#> 140 69.5 0.4388983 0.4388983 0.1413973 0.1413973 0.419704442 0.419704442
#> 141 70.0 0.4376347 0.4376347 0.1405219 0.1405219 0.421843379 0.421843379
#> 142 70.5 0.4363857 0.4363857 0.1396476 0.1396476 0.423966671 0.423966671
#>  [ reached 'max' / getOption("max.print") -- omitted 19859 rows ]
#> 
#> attr(,"class")
#> [1] "sir_subpops_model"
# }