Estimate seroprevalence and foi by combining mixture model and regression

Refers to section 11.2 - 11.4

Usage

estimate_from_mixture(
  age,
  antibody_level,
  threshold_status = NULL,
  mixture_model,
  s = "ps",
  sp = 83,
  monotonize = TRUE
)

Arguments

age

- vector of age

antibody_level

- vector of the corresponding raw antibody level

threshold_status

- sero status using threshold approach in line listing (optional, for visualization and comparison only)

mixture_model

- mixture_model object generated by serosv::mixture_model()

s

- smoothing basis used to fit antibody level

sp

- smoothing parameter

monotonize

- whether to monotonize seroprevalence (default to TRUE)

Value

a list of class estimated_from_mixture with the following items

df

the dataframe used for fitting the model

info

a fitted "gam" model for mu(a)

sp

seroprevalence

foi

force of infection

threshold_status

serostatus using threshold method only if provided

See also